refactor(Stack): move stack to a specific folder in stacks
