+++ stacks/IStack.ts
